+++ app/src/main/java/gateway/controller/CtrlFileUpload.java
@@ -67,8 +67,7 @@ public class CtrlFileUpload
 
 		try {
 			IWorkerService server = ServiceWorker.getServer ();
-			UploadFileArgs queryUpload =
-				new UploadFileArgs (s.fileUUID.toString (), args.fileContent);
+			UploadFileArgs queryUpload = new UploadFileArgs (s.fileUUID, args.fileContent);
 			server.uploadFile (queryUpload);
 
 			s.code = 201;
@@ -76,7 +75,7 @@ public class CtrlFileUpload
 			s.msg = "Your file is being uploaded";
 		} catch (Exception e) {
 			System.err.println ("Can't connect to RMI");
-			e.printStackTrace ();
+			System.err.println (e);
 
 			s.code = 500;
 			s.error = true;